refactor(gitops_server): Started a better folder structure
