+++ gitops_server/workers/__init__.py
@@ -1,0 +1,2 @@
+from .deployer import DeployQueueWorker  # NOQA
+from .status_updater import DeploymentStatusWorker  # NOQA
