+++ gitops_server/workers/deployer/__init__.py
@@ -1,0 +1,1 @@
+from .worker import DeployQueueWorker  # NOQA
