+++ gitops_server/workers/status_updater/__init__.py
@@ -1,0 +1,1 @@
+from .worker import DeploymentStatusWorker  # NOQA
